# docs/swagger.yaml
basePath: /api
definitions:
  controllers.AuthorRequestPayload:
    properties:
      accountNumber:
        type: string
      bankId:
        type: integer
      instagram:
        type: string
      penName:
        type: string
      phone:
        type: string
    type: object
  controllers.AuthorStatusResponse:
    properties:
      isAuthor:
        type: boolean
      user:
        $ref: '#/definitions/tables.User'
    type: object
  controllers.BankResponse:
    properties:
      bankCode:
        type: string
      bankId:
        type: integer
      bankName:
        type: string
    type: object
  controllers.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  controllers.LoginRequest:
    properties:
      password:
        example: password123
        type: string
      username:
        example: testuser
        type: string
    type: object
  controllers.LoginSuccessResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/tables.User'
    type: object
  fiber.Map:
    additionalProperties: true
    type: object
  tables.User:
    properties:
      accountNumber:
        type: string
      avatarUrl:
        type: string
      bankId:
        type: integer
      createDatetime:
        type: string
      email:
        type: string
      flgAuthor:
        type: string
      fullName:
        type: string
      instagram:
        type: string
      isEmailVerified:
        type: boolean
      loginWith:
        type: string
      password:
        type: string
      penName:
        type: string
      phone:
        type: string
      updateDatetime:
        type: string
      userCode:
        type: string
      userId:
        type: integer
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@noversystem.dev
    name: API Support
  description: Ini adalah dokumentasi API untuk Nover System.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Nover System API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Mengotentikasi pengguna dengan username dan password, lalu memberikan
        JWT.
      parameters:
      - description: Kredensial Login dengan Username
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.LoginSuccessResponse'
        "400":
          description: Input tidak valid
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Kredensial tidak valid
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Error internal server
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Login pengguna
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Membuat akun pengguna baru dengan email, username, dan password.
      parameters:
      - description: Informasi Registrasi Pengguna
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/tables.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/tables.User'
        "400":
          description: Input tidak valid
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "409":
          description: Email atau Username sudah terdaftar
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Error internal server
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Registrasi pengguna baru
      tags:
      - Authentication
  /v1/bank/get:
    get:
      consumes:
      - application/json
      description: Mengambil daftar semua bank yang aktif, diurutkan berdasarkan nama.
      produces:
      - application/json
      responses:
        "200":
          description: Daftar bank yang berhasil diambil
          schema:
            items:
              $ref: '#/definitions/controllers.BankResponse'
            type: array
        "500":
          description: Terjadi kesalahan internal pada server
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Daftar Bank
      tags:
      - Bank
  /v1/user/author-status:
    get:
      description: Memvalidasi token, memeriksa apakah pengguna adalah penulis, dan
        mengembalikan data profil lengkap pengguna.
      produces:
      - application/json
      responses:
        "200":
          description: Status penulis dan data profil lengkap
          schema:
            $ref: '#/definitions/controllers.AuthorStatusResponse'
        "401":
          description: Tidak terotentikasi
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Pengguna tidak ditemukan
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Error internal server
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Cek Status Penulis & Profil
      tags:
      - User
  /v1/user/request-author:
    post:
      consumes:
      - application/json
      description: Memperbarui profil pengguna untuk menjadi penulis dengan melengkapi
        data yang diperlukan. Endpoint ini memerlukan otentikasi.
      parameters:
      - description: Data untuk menjadi penulis
        in: body
        name: author_request
        required: true
        schema:
          $ref: '#/definitions/controllers.AuthorRequestPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Pesan sukses
          schema:
            properties:
              code:
                type: string
              message:
                type: string
            type: object
        "400":
          description: Input tidak valid
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Tidak terotentikasi
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "409":
          description: Nama pena sudah digunakan
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Error internal server
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Permintaan menjadi Penulis
      tags:
      - User
swagger: "2.0"